useIsMobileMode hook added to detect when the screen is in mobile mode
