+++ src/components/CategorieSideBar/index.jsx
@@ -5,14 +5,15 @@ import QueryFilterContext from "../../context/filtersContext";
 import CategoryIcon from "../../assets/category-icon.svg";
 import CategoriesList from "./CategoriesList";
 import { Dialog } from "primereact/dialog";
+import { useIsMobileMode } from "../../hooks/useIsMobileMode";
 import "./index.css";
 
-function CategorieSideBar(forceMobileMode = false) {
+function CategorieSideBar({forceMobileMode = false}) {
+  const {mobileMode} = useIsMobileMode({forceMobileMode:forceMobileMode})
   const [categories, setCategories] = useState([]);
   const [activeCategory, setActiveCategory] = useState(null);
   const [loading, setLoading] = useState(false);
   const { setFilter, getActiveFilter } = useContext(QueryFilterContext);
-  const [mobileMode, setMobileMode] = useState(window.innerWidth > 830 && forceMobileMode === false?false:true);
   const [showModal, setShowModal] = useState(false);
 
   useEffect(() => {
@@ -23,24 +24,6 @@ function CategorieSideBar(forceMobileMode = false) {
       setActiveCategory(getActiveFilter("categoria"));
     });
   }, []);
-
-  //effect to capture the window's width
-  useEffect(() => {
-    if(forceMobileMode === false){
-      const handleResize = () => {
-        if(window.innerWidth <= 830 && !mobileMode){
-          setMobileMode(true);
-        }
-        if(window.innerWidth > 830 && mobileMode){
-          setMobileMode(false);
-        }
-      };
-      window.addEventListener("resize", handleResize);
-      return () => {
-        window.removeEventListener("resize", handleResize);
-      };
-    }
-  });
 
   function handleSetActiveCategory(category) {
     setActiveCategory(category);
